refactor: Move discount card and product ordering card to separate files
