+++ apps/dashboard/src/forms/fr_accounting/product.page.tsx
@@ -1,195 +1,11 @@
-import { Badge } from "@/components/ui/badge"
-import {
-    Card,
-    CardDescription,
-    CardFooter,
-    CardHeader,
-    CardTitle
-} from "@/components/ui/card"
-import { Input } from "@/components/ui/input"
-import { Switch } from "@/components/ui/switch"
-import {
-    Tooltip,
-    TooltipContent,
-    TooltipTrigger
-} from "@/components/ui/tooltip"
-import {
-    belongsToSection,
-    getProductWithAdjustedPrice
-} from "@/forms/fr_accounting/accounting_utilities"
-import { useAccountingMutation } from "@/forms/fr_accounting/useAccounting"
-import { useDashboard } from "@/shared/hooks/api/useDashboard"
-import { useOrders } from "@/shared/hooks/api/useOrders"
+import { belongsToSection } from "@/forms/fr_accounting/accounting_utilities"
+import { DiscountCard } from "@/forms/fr_accounting/components/DiscountCard"
+import { ProductOrderingCard } from "@/forms/fr_accounting/components/ProductOrderingCard"
 import { useProducts } from "@/shared/hooks/api/useProducts"
 import { RegistrationSection } from "@/shared/vars"
 import React, { useMemo } from "react"
 import { Category, Product } from "../../store/products/products_slice"
-import { cx } from "../../utils/cx"
-import { formatCurrency } from "../../utils/format_currency"
 import { FormWrapper } from "../FormWrapper"
-
-function InputCard({ product }: { product: Product }) {
-    const { updateCache } = useDashboard()
-    const { data: orders } = useOrders()
-    const { data: products } = useProducts()
-
-    const { setProductOrder } = useAccountingMutation({
-        onSuccess: updateCache
-    })
-
-    const order = orders.find(x => x.product.id === product.id)
-    const productWithAdjustedPrice = getProductWithAdjustedPrice(
-        product.id,
-        orders,
-        products
-    )
-    const productAdjustedUnitPrice =
-        (productWithAdjustedPrice?.adjustedPrice ?? 0) / (order?.quantity ?? 1)
-
-    const selected = orders.find(x => x.product.id === product.id)
-
-    const productPackage = getProductWithAdjustedPrice(
-        order?.product.id,
-        orders,
-        products
-    )
-
-    const packageProductBaseQuantity =
-        productPackage?.child_products?.find(
-            current => current.child_product.id === product.id
-        )?.quantity ?? 0
-
-    async function onChange(quantity: number) {
-        if (isNaN(quantity)) quantity = 0
-        if (product.max_quantity != null && quantity > product.max_quantity)
-            quantity = product.max_quantity
-        if (quantity <= 0 || quantity == null) {
-            await setProductOrder(product.id, 0)
-        } else if (
-            !(
-                // If the product has a max of 1 and is already
-                // included, don't allow it to be added or incremented
-                (
-                    product.max_quantity != null &&
-                    product.max_quantity <= 1 &&
-                    packageProductBaseQuantity > 0
-                )
-            )
-        ) {
-            await setProductOrder(product.id, quantity)
-        }
-    }
-
-    const packageName =
-        productPackage?.short_name || productPackage?.name || "package"
-
-    return (
-        <Card
-            className={cx("w-full", {
-                "border-2 border-emerald-500": selected
-            })}
-        >
-            <CardHeader className="flex flex-row justify-between">
-                <div className="">
-                    <CardTitle className="text-xl">
-                        {product.short_name || product.name}
-                    </CardTitle>
-                    <CardDescription>{product.description}</CardDescription>
-                </div>
-                <div className="!mt-0 flex flex-1 items-start justify-end">
-                    {product.max_quantity <= 1 ? (
-                        <Tooltip>
-                            <TooltipTrigger>
-                                <Switch
-                                    onCheckedChange={value =>
-                                        onChange(value ? 1 : 0)
-                                    }
-                                    checked={
-                                        selected != null ||
-                                        packageProductBaseQuantity > 0
-                                    }
-                                    disabled={packageProductBaseQuantity > 0}
-                                />
-                            </TooltipTrigger>
-                            <TooltipContent
-                                hidden={packageProductBaseQuantity <= 0}
-                            >
-                                Included in package
-                            </TooltipContent>
-                        </Tooltip>
-                    ) : (
-                        <Input
-                            placeholder="Quantity"
-                            className="w-40"
-                            value={selected?.quantity.toString() ?? "0"}
-                            min={0}
-                            max={product.max_quantity}
-                            type="number"
-                            onChange={event =>
-                                onChange(parseInt(event.target.value))
-                            }
-                        />
-                    )}
-                </div>
-            </CardHeader>
-            <CardFooter>
-                <div className="flex items-center gap-3">
-                    {selected != null && selected.quantity > 1 && (
-                        <Badge className="rounded p-2" variant={"outline"}>
-                            Unit price{" "}
-                            {formatCurrency(productAdjustedUnitPrice)} kr
-                        </Badge>
-                    )}
-                    <p className="rounded bg-emerald-400 p-1 px-3 text-lg text-white">
-                        {product.max_quantity <= 1 &&
-                        packageProductBaseQuantity > 0
-                            ? `Included In ${packageName}`
-                            : packageProductBaseQuantity > 0 &&
-                              (selected?.quantity ?? 0) <= 0
-                            ? "0 kr"
-                            : `${formatCurrency(
-                                  productWithAdjustedPrice?.adjustedPrice ?? NaN
-                              )} kr`}
-                    </p>
-                    {product.max_quantity > 1 &&
-                        packageProductBaseQuantity > 0 && (
-                            <>
-                                <p className="text-slate-500">in addition to</p>
-                                <p className="rounded bg-emerald-400 p-1 px-3 text-sm text-white">
-                                    {packageProductBaseQuantity} included in{" "}
-                                    {packageName}
-                                </p>
-                            </>
-                        )}
-                </div>
-            </CardFooter>
-        </Card>
-    )
-}
-
-export function DiscountCard({ product }: { product: Product }) {
-    return (
-        <div className="flex items-center justify-center">
-            <div className="w-full rounded bg-gradient-to-r from-yellow-500 to-yellow-300 p-1">
-                <div className="flex h-full w-full bg-white p-5">
-                    <div className="flex-[2]">
-                        <h3 className="bg-white text-lg text-yellow-600">
-                            {product.short_name || product.name}
-                        </h3>
-                        <p className="text-xs text-yellow-800">
-                            {product.description}
-                        </p>
-                    </div>
-                    <div className="flex flex-1 items-center justify-center">
-                        <p className="rounded bg-gradient-to-br from-yellow-600 to-yellow-500 p-2 px-4 font-bold text-white">
-                            {formatCurrency(product.unit_price)} kr
-                        </p>
-                    </div>
-                </div>
-            </div>
-        </div>
-    )
-}
 
 export function ProductFormPage({ section }: { section: RegistrationSection }) {
     const { data: allProducts } = useProducts()
@@ -273,7 +89,7 @@ export function ProductFormPage({ section }: { section: RegistrationSection }) {
                                             current => current.unit_price >= 0
                                         )
                                         .map(current => (
-                                            <InputCard
+                                            <ProductOrderingCard
                                                 key={current.id}
                                                 product={current}
                                             />
